requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
PyJWT==2.10.1
bcrypt==4.1.3
passlib==1.7.4
email-validator==2.3.0
python-multipart==0.0.20
dnspython==2.8.0
starlette==0.37.2
anyio==4.11.0
h11==0.16.0
typing_extensions==4.15.0

